% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{ApproximateProgress}
\alias{ApproximateProgress}
\title{ApproximateProgress Object}
\usage{
ApproximateProgress(position = NULL, percentComplete = NULL,
  remainingTime = NULL)
}
\arguments{
\item{position}{Obsolete}

\item{percentComplete}{Obsolete}

\item{remainingTime}{Obsolete}
}
\value{
ApproximateProgress object
}
\description{
ApproximateProgress Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Obsolete in favor of ApproximateReportedProgress and ApproximateSplitRequest.
}

