% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{ApproximateSplitRequest}
\alias{ApproximateSplitRequest}
\title{ApproximateSplitRequest Object}
\usage{
ApproximateSplitRequest(position = NULL, fractionConsumed = NULL)
}
\arguments{
\item{position}{A Position at which to split the work item}

\item{fractionConsumed}{A fraction at which to split the work item, from 0}
}
\value{
ApproximateSplitRequest object
}
\description{
ApproximateSplitRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A suggestion by the service to the worker to dynamically split the WorkItem.
}

