% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{AutoscalingSettings}
\alias{AutoscalingSettings}
\title{AutoscalingSettings Object}
\usage{
AutoscalingSettings(algorithm = NULL, maxNumWorkers = NULL)
}
\arguments{
\item{algorithm}{The algorithm to use for autoscaling}

\item{maxNumWorkers}{The maximum number of workers to cap scaling at}
}
\value{
AutoscalingSettings object
}
\description{
AutoscalingSettings Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Settings for WorkerPool autoscaling.
}

