% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{ComponentSource}
\alias{ComponentSource}
\title{ComponentSource Object}
\usage{
ComponentSource(originalTransformOrCollection = NULL, name = NULL,
  userName = NULL)
}
\arguments{
\item{originalTransformOrCollection}{User name for the original user transform or collection with which this}

\item{name}{Dataflow service generated name for this source}

\item{userName}{Human-readable name for this transform; may be user or system generated}
}
\value{
ComponentSource object
}
\description{
ComponentSource Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Description of an interstitial value between transforms in an executionstage.
}
