% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{ComponentTransform}
\alias{ComponentTransform}
\title{ComponentTransform Object}
\usage{
ComponentTransform(originalTransform = NULL, name = NULL, userName = NULL)
}
\arguments{
\item{originalTransform}{User name for the original user transform with which this transform is}

\item{name}{Dataflow service generated name for this source}

\item{userName}{Human-readable name for this transform; may be user or system generated}
}
\value{
ComponentTransform object
}
\description{
ComponentTransform Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Description of a transform executed as part of an execution stage.
}
