% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{ComputationTopology}
\alias{ComputationTopology}
\title{ComputationTopology Object}
\usage{
ComputationTopology(outputs = NULL, stateFamilies = NULL,
  systemStageName = NULL, computationId = NULL, inputs = NULL,
  keyRanges = NULL, userStageName = NULL)
}
\arguments{
\item{outputs}{The outputs from the computation}

\item{stateFamilies}{The state family values}

\item{systemStageName}{The system stage name}

\item{computationId}{The ID of the computation}

\item{inputs}{The inputs to the computation}

\item{keyRanges}{The key ranges processed by the computation}

\item{userStageName}{The user stage name}
}
\value{
ComputationTopology object
}
\description{
ComputationTopology Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
All configuration data for a particular Computation.
}
