% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{CounterMetadata}
\alias{CounterMetadata}
\title{CounterMetadata Object}
\usage{
CounterMetadata(description = NULL, standardUnits = NULL,
  otherUnits = NULL)
}
\arguments{
\item{description}{Human-readable description of the counter semantics}

\item{standardUnits}{System defined Units, see above enum}

\item{otherUnits}{A string referring to the unit type}
}
\value{
CounterMetadata object
}
\description{
CounterMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
CounterMetadata includes all static non-name non-value counter attributes.
}

