% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{CounterStructuredName}
\alias{CounterStructuredName}
\title{CounterStructuredName Object}
\usage{
CounterStructuredName(name = NULL, standardOrigin = NULL,
  otherOrigin = NULL, originalStepName = NULL, componentStepName = NULL,
  executionStepName = NULL, workerId = NULL, portion = NULL)
}
\arguments{
\item{name}{Counter name}

\item{standardOrigin}{One of the standard Origins defined above}

\item{otherOrigin}{A string containing the origin of the counter}

\item{originalStepName}{System generated name of the original step in the user's graph, before optimization}

\item{componentStepName}{Name of the optimized step being executed by the workers}

\item{executionStepName}{Name of the stage}

\item{workerId}{ID of a particular worker}

\item{portion}{Portion of this counter, either key or value}
}
\value{
CounterStructuredName object
}
\description{
CounterStructuredName Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Identifies a counter within a per-job namespace. Counters whose structured names are the same get merged into a single value for the job.
}

