% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{CounterStructuredNameAndMetadata}
\alias{CounterStructuredNameAndMetadata}
\title{CounterStructuredNameAndMetadata Object}
\usage{
CounterStructuredNameAndMetadata(name = NULL, metadata = NULL)
}
\arguments{
\item{name}{Structured name of the counter}

\item{metadata}{Metadata associated with a counter}
}
\value{
CounterStructuredNameAndMetadata object
}
\description{
CounterStructuredNameAndMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A single message which encapsulates structured name and metadata for a given counter.
}

