% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{CounterUpdate}
\alias{CounterUpdate}
\title{CounterUpdate Object}
\usage{
CounterUpdate(nameAndKind = NULL, shortId = NULL,
  structuredNameAndMetadata = NULL, cumulative = NULL, integer = NULL,
  floatingPoint = NULL, boolean = NULL, integerMean = NULL,
  floatingPointMean = NULL, integerList = NULL, floatingPointList = NULL,
  stringList = NULL, internal = NULL)
}
\arguments{
\item{nameAndKind}{Counter name and aggregation type}

\item{shortId}{The service-generated short identifier for this counter}

\item{structuredNameAndMetadata}{Counter structured name and metadata}

\item{cumulative}{True if this counter is reported as the total cumulative aggregate value accumulated since the worker started working on this WorkItem}

\item{integer}{Integer value for Sum, Max, Min}

\item{floatingPoint}{Floating point value for Sum, Max, Min}

\item{boolean}{Boolean value for And, Or}

\item{integerMean}{Integer mean aggregation value for Mean}

\item{floatingPointMean}{Floating point mean aggregation value for Mean}

\item{integerList}{List of integers, for Set}

\item{floatingPointList}{List of floating point numbers, for Set}

\item{stringList}{List of strings, for Set}

\item{internal}{Value for internally-defined counters used by the Dataflow service}
}
\value{
CounterUpdate object
}
\description{
CounterUpdate Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An update to a Counter sent from a worker.
}

