% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{CPUTime}
\alias{CPUTime}
\title{CPUTime Object}
\usage{
CPUTime(rate = NULL, timestamp = NULL, totalMs = NULL)
}
\arguments{
\item{rate}{Average CPU utilization rate (% non-idle cpu / second) since previous}

\item{timestamp}{Timestamp of the measurement}

\item{totalMs}{Total active CPU time across all cores (ie}
}
\value{
CPUTime object
}
\description{
CPUTime Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Modeled after information exposed by /proc/stat.
}
