% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{CreateJobFromTemplateRequest}
\alias{CreateJobFromTemplateRequest}
\title{CreateJobFromTemplateRequest Object}
\usage{
CreateJobFromTemplateRequest(CreateJobFromTemplateRequest.parameters = NULL,
  gcsPath = NULL, parameters = NULL)
}
\arguments{
\item{CreateJobFromTemplateRequest.parameters}{The \link{CreateJobFromTemplateRequest.parameters} object or list of objects}

\item{gcsPath}{A path to the serialized JSON representation of the job}

\item{parameters}{Dynamic parameterization of the job's runtime environment}
}
\value{
CreateJobFromTemplateRequest object
}
\description{
CreateJobFromTemplateRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request to create a Dataflow job.
}
\seealso{
Other CreateJobFromTemplateRequest functions: \code{\link{CreateJobFromTemplateRequest.parameters}}
}

