% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{CustomSourceLocation}
\alias{CustomSourceLocation}
\title{CustomSourceLocation Object}
\usage{
CustomSourceLocation(stateful = NULL)
}
\arguments{
\item{stateful}{Whether this source is stateful}
}
\value{
CustomSourceLocation object
}
\description{
CustomSourceLocation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Identifies the location of a custom souce.
}
