% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{DataDiskAssignment}
\alias{DataDiskAssignment}
\title{DataDiskAssignment Object}
\usage{
DataDiskAssignment(vmInstance = NULL, dataDisks = NULL)
}
\arguments{
\item{vmInstance}{VM instance name the data disks mounted to, for example 'myproject-1014-104817-4c2-harness-0'}

\item{dataDisks}{Mounted data disks}
}
\value{
DataDiskAssignment object
}
\description{
DataDiskAssignment Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Data disk assignment for a given VM instance.
}

