% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_functions.R
\docType{package}
\name{dataflow_googleAuthR}
\alias{dataflow_googleAuthR}
\alias{dataflow_googleAuthR-package}
\title{Google Dataflow API
Manages Google Cloud Dataflow projects on Google Cloud Platform.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2017-03-05 19:41:42
filename: /Users/mark/dev/R/autoGoogleAPI/googledataflowv1b3.auto/R/dataflow_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/userinfo.email
}
}
