% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{DerivedSource}
\alias{DerivedSource}
\title{DerivedSource Object}
\usage{
DerivedSource(source = NULL, derivationMode = NULL)
}
\arguments{
\item{source}{Specification of the source}

\item{derivationMode}{What source to base the produced source on (if any)}
}
\value{
DerivedSource object
}
\description{
DerivedSource Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Specification of one of the bundles produced as a result of splittinga Source (e.g. when executing a SourceSplitRequest, or whensplitting an active task using WorkItemStatus.dynamic_source_split),relative to the source being split.
}
