% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{Disk}
\alias{Disk}
\title{Disk Object}
\usage{
Disk(sizeGb = NULL, diskType = NULL, mountPoint = NULL)
}
\arguments{
\item{sizeGb}{Size of disk in GB}

\item{diskType}{Disk storage type, as defined by Google Compute Engine}

\item{mountPoint}{Directory in a VM where disk is mounted}
}
\value{
Disk object
}
\description{
Disk Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Describes the data disk used by a workflow job.
}

