% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{DisplayData}
\alias{DisplayData}
\title{DisplayData Object}
\usage{
DisplayData(strValue = NULL, durationValue = NULL, int64Value = NULL,
  namespace = NULL, floatValue = NULL, key = NULL, shortStrValue = NULL,
  label = NULL, url = NULL, timestampValue = NULL, boolValue = NULL,
  javaClassValue = NULL)
}
\arguments{
\item{strValue}{Contains value if the data is of string type}

\item{durationValue}{Contains value if the data is of duration type}

\item{int64Value}{Contains value if the data is of int64 type}

\item{namespace}{The namespace for the key}

\item{floatValue}{Contains value if the data is of float type}

\item{key}{The key identifying the display data}

\item{shortStrValue}{A possible additional shorter value to display}

\item{label}{An optional label to display in a dax UI for the element}

\item{url}{An optional full URL}

\item{timestampValue}{Contains value if the data is of timestamp type}

\item{boolValue}{Contains value if the data is of a boolean type}

\item{javaClassValue}{Contains value if the data is of java class type}
}
\value{
DisplayData object
}
\description{
DisplayData Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Data provided with a pipeline or transform to provide descriptive info.
}
