% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{DistributionUpdate}
\alias{DistributionUpdate}
\title{DistributionUpdate Object}
\usage{
DistributionUpdate(count = NULL, min = NULL, sumOfSquares = NULL,
  max = NULL, sum = NULL)
}
\arguments{
\item{count}{The count of the number of elements present in the distribution}

\item{min}{The minimum value present in the distribution}

\item{sumOfSquares}{Use a double since the sum of squares is likely to overflow int64}

\item{max}{The maximum value present in the distribution}

\item{sum}{Use an int64 since we'd prefer the added precision}
}
\value{
DistributionUpdate object
}
\description{
DistributionUpdate Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A metric value representing a distribution.
}
