% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{DynamicSourceSplit}
\alias{DynamicSourceSplit}
\title{DynamicSourceSplit Object}
\usage{
DynamicSourceSplit(primary = NULL, residual = NULL)
}
\arguments{
\item{primary}{Primary part (continued to be processed by worker)}

\item{residual}{Residual part (returned to the pool of work)}
}
\value{
DynamicSourceSplit object
}
\description{
DynamicSourceSplit Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
When a task splits using WorkItemStatus.dynamic_source_split, thismessage describes the two parts of the split relative to thedescription of the current task's input.
}
