% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{Environment.userAgent}
\alias{Environment.userAgent}
\title{Environment.userAgent Object}
\usage{
Environment.userAgent()
}
\value{
Environment.userAgent object
}
\description{
Environment.userAgent Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A description of the process that generated the request.
}
\seealso{
Other Environment functions: \code{\link{Environment.internalExperiments}},
  \code{\link{Environment.sdkPipelineOptions}},
  \code{\link{Environment.version}},
  \code{\link{Environment}}
}

