% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{Environment.version}
\alias{Environment.version}
\title{Environment.version Object}
\usage{
Environment.version()
}
\value{
Environment.version object
}
\description{
Environment.version Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A structure describing which components and their versions of the serviceare required in order to run the job.
}
\seealso{
Other Environment functions: \code{\link{Environment.internalExperiments}},
  \code{\link{Environment.sdkPipelineOptions}},
  \code{\link{Environment.userAgent}},
  \code{\link{Environment}}
}
