% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{ExecutionStageSummary}
\alias{ExecutionStageSummary}
\title{ExecutionStageSummary Object}
\usage{
ExecutionStageSummary(inputSource = NULL, id = NULL,
  componentTransform = NULL, componentSource = NULL, outputSource = NULL,
  name = NULL)
}
\arguments{
\item{inputSource}{Input sources for this stage}

\item{id}{Dataflow service generated id for this stage}

\item{componentTransform}{Transforms that comprise this execution stage}

\item{componentSource}{Collections produced and consumed by component transforms of this stage}

\item{outputSource}{Output sources for this stage}

\item{name}{Dataflow service generated name for this stage}
}
\value{
ExecutionStageSummary object
}
\description{
ExecutionStageSummary Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Description of the composing transforms, names/ids, and input/outputs of astage of execution.  Some composing transforms and sources may have beengenerated by the Dataflow service during execution planning.
}
