% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{FailedLocation}
\alias{FailedLocation}
\title{FailedLocation Object}
\usage{
FailedLocation(name = NULL)
}
\arguments{
\item{name}{The name of the failed location}
}
\value{
FailedLocation object
}
\description{
FailedLocation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Indicates which location failed to respond to a request for data.
}
