% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{FlattenInstruction}
\alias{FlattenInstruction}
\title{FlattenInstruction Object}
\usage{
FlattenInstruction(inputs = NULL)
}
\arguments{
\item{inputs}{Describes the inputs to the flatten instruction}
}
\value{
FlattenInstruction object
}
\description{
FlattenInstruction Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An instruction that copies its inputs (zero or more) to its (single) output.
}
