% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{FloatingPointList}
\alias{FloatingPointList}
\title{FloatingPointList Object}
\usage{
FloatingPointList(elements = NULL)
}
\arguments{
\item{elements}{Elements of the list}
}
\value{
FloatingPointList object
}
\description{
FloatingPointList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A metric value representing a list of floating point numbers.
}

