% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{FloatingPointMean}
\alias{FloatingPointMean}
\title{FloatingPointMean Object}
\usage{
FloatingPointMean(sum = NULL, count = NULL)
}
\arguments{
\item{sum}{The sum of all values being aggregated}

\item{count}{The number of values being aggregated}
}
\value{
FloatingPointMean object
}
\description{
FloatingPointMean Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A representation of a floating point mean metric contribution.
}

