% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{GetDebugConfigRequest}
\alias{GetDebugConfigRequest}
\title{Google Dataflow API Objects 
Develops and executes data processing patterns like ETL, batch computation, and continuous computation.}
\usage{
GetDebugConfigRequest(workerId = NULL, componentId = NULL)
}
\arguments{
\item{workerId}{The worker id, i}

\item{componentId}{The internal component id for which debug configuration is requested}
}
\value{
GetDebugConfigRequest object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-03 23:18:12
filename: /Users/mark/dev/R/autoGoogleAPI/googledataflowv1b3.auto/R/dataflow_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
GetDebugConfigRequest Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request to get updated debug configuration for component.
}

