% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{GetDebugConfigResponse}
\alias{GetDebugConfigResponse}
\title{GetDebugConfigResponse Object}
\usage{
GetDebugConfigResponse(config = NULL)
}
\arguments{
\item{config}{The encoded debug configuration for the requested component}
}
\value{
GetDebugConfigResponse object
}
\description{
GetDebugConfigResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response to a get debug configuration request.
}
