% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{InstructionInput}
\alias{InstructionInput}
\title{InstructionInput Object}
\usage{
InstructionInput(producerInstructionIndex = NULL, outputNum = NULL)
}
\arguments{
\item{producerInstructionIndex}{The index (origin zero) of the parallel instruction that produces the output to be consumed by this input}

\item{outputNum}{The output index (origin zero) within the producer}
}
\value{
InstructionInput object
}
\description{
InstructionInput Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An input of an instruction, as a reference to an output of a producer instruction.
}

