% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{InstructionOutput.codec}
\alias{InstructionOutput.codec}
\title{InstructionOutput.codec Object}
\usage{
InstructionOutput.codec()
}
\value{
InstructionOutput.codec object
}
\description{
InstructionOutput.codec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The codec to use to encode data being written via this output.
}
\seealso{
Other InstructionOutput functions: \code{\link{InstructionOutput}}
}

