% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{InstructionOutput}
\alias{InstructionOutput}
\title{InstructionOutput Object}
\usage{
InstructionOutput(InstructionOutput.codec = NULL, name = NULL,
  systemName = NULL, originalName = NULL, codec = NULL,
  onlyCountKeyBytes = NULL, onlyCountValueBytes = NULL)
}
\arguments{
\item{InstructionOutput.codec}{The \link{InstructionOutput.codec} object or list of objects}

\item{name}{The user-provided name of this output}

\item{systemName}{System-defined name of this output}

\item{originalName}{System-defined name for this output in the original workflow graph}

\item{codec}{The codec to use to encode data being written via this output}

\item{onlyCountKeyBytes}{For system-generated byte and mean byte metrics, certain instructions should only report the key size}

\item{onlyCountValueBytes}{For system-generated byte and mean byte metrics, certain instructions should only report the value size}
}
\value{
InstructionOutput object
}
\description{
InstructionOutput Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An output of an instruction.
}
\seealso{
Other InstructionOutput functions: \code{\link{InstructionOutput.codec}}
}

