% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{IntegerList}
\alias{IntegerList}
\title{IntegerList Object}
\usage{
IntegerList(elements = NULL)
}
\arguments{
\item{elements}{Elements of the list}
}
\value{
IntegerList object
}
\description{
IntegerList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A metric value representing a list of integers.
}

