% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{IntegerMean}
\alias{IntegerMean}
\title{IntegerMean Object}
\usage{
IntegerMean(sum = NULL, count = NULL)
}
\arguments{
\item{sum}{The sum of all values being aggregated}

\item{count}{The number of values being aggregated}
}
\value{
IntegerMean object
}
\description{
IntegerMean Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A representation of an integer mean metric contribution.
}
