% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{Job.labels}
\alias{Job.labels}
\title{Job.labels Object}
\usage{
Job.labels()
}
\value{
Job.labels object
}
\description{
Job.labels Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
User-defined labels for this job.The labels map can contain no more than 64 entries.  Entries of the labelsmap are UTF8 strings that comply with the following restrictions:* Keys must conform to regexp:  \p{Ll}\p{Lo}{0,62}* Values must conform to regexp:  [\p{Ll}\p{Lo}\p{N}_-]{0,63}* Both keys and values are additionally constrained to be <= 128 bytes insize.
}
\seealso{
Other Job functions: \code{\link{Job.transformNameMapping}},
  \code{\link{Job}}
}
