% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{Job}
\alias{Job}
\title{Job Object}
\usage{
Job(Job.transformNameMapping = NULL, Job.labels = NULL, id = NULL,
  projectId = NULL, name = NULL, type = NULL, environment = NULL,
  steps = NULL, currentState = NULL, currentStateTime = NULL,
  requestedState = NULL, executionInfo = NULL, createTime = NULL,
  replaceJobId = NULL, transformNameMapping = NULL,
  clientRequestId = NULL, replacedByJobId = NULL, tempFiles = NULL,
  labels = NULL)
}
\arguments{
\item{Job.transformNameMapping}{The \link{Job.transformNameMapping} object or list of objects}

\item{Job.labels}{The \link{Job.labels} object or list of objects}

\item{id}{The unique ID of this job}

\item{projectId}{The project which owns the job}

\item{name}{The user-specified Dataflow job name}

\item{type}{The type of dataflow job}

\item{environment}{Environment for the job}

\item{steps}{The top-level steps that constitute the entire job}

\item{currentState}{The current state of the job}

\item{currentStateTime}{The timestamp associated with the current state}

\item{requestedState}{The job's requested state}

\item{executionInfo}{Information about how the Dataflow service will actually run the job}

\item{createTime}{Timestamp when job was initially created}

\item{replaceJobId}{If this job is an update of an existing job, this field will be the ID of the job it replaced}

\item{transformNameMapping}{Map of transform name prefixes of the job to be replaced to the corresponding name prefixes of the new job}

\item{clientRequestId}{Client's unique identifier of the job, re-used by SDK across retried attempts}

\item{replacedByJobId}{If another job is an update of this job (and thus, this job is in JOB_STATE_UPDATED), this field will contain the ID of that job}

\item{tempFiles}{A set of files the system should be aware of that are used for temporary storage}

\item{labels}{User-defined labels for this job}
}
\value{
Job object
}
\description{
Job Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Defines a job to be run by the Dataflow service.
}
\seealso{
Other Job functions: \code{\link{Job.labels}},
  \code{\link{Job.transformNameMapping}}
}

