% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{Job.transformNameMapping}
\alias{Job.transformNameMapping}
\title{Job.transformNameMapping Object}
\usage{
Job.transformNameMapping()
}
\value{
Job.transformNameMapping object
}
\description{
Job.transformNameMapping Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The map of transform name prefixes of the job to be replaced to thecorresponding name prefixes of the new job.
}
\seealso{
Other Job functions: \code{\link{Job.labels}},
  \code{\link{Job}}
}
