% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{JobExecutionInfo}
\alias{JobExecutionInfo}
\title{JobExecutionInfo Object}
\usage{
JobExecutionInfo(JobExecutionInfo.stages = NULL, stages = NULL)
}
\arguments{
\item{JobExecutionInfo.stages}{The \link{JobExecutionInfo.stages} object or list of objects}

\item{stages}{A mapping from each stage to the information about that stage}
}
\value{
JobExecutionInfo object
}
\description{
JobExecutionInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Additional information about how a Cloud Dataflow job will be executed thatisn't contained in the submitted job.
}
\seealso{
Other JobExecutionInfo functions: \code{\link{JobExecutionInfo.stages}}
}
