% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{JobExecutionInfo.stages}
\alias{JobExecutionInfo.stages}
\title{JobExecutionInfo.stages Object}
\usage{
JobExecutionInfo.stages()
}
\value{
JobExecutionInfo.stages object
}
\description{
JobExecutionInfo.stages Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A mapping from each stage to the information about that stage.
}
\seealso{
Other JobExecutionInfo functions: \code{\link{JobExecutionInfo}}
}

