% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{JobExecutionStageInfo}
\alias{JobExecutionStageInfo}
\title{JobExecutionStageInfo Object}
\usage{
JobExecutionStageInfo(stepName = NULL)
}
\arguments{
\item{stepName}{The steps associated with the execution stage}
}
\value{
JobExecutionStageInfo object
}
\description{
JobExecutionStageInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains information about how a particulargoogle.dataflow.v1beta3.Step will be executed.
}
