% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{JobMessage}
\alias{JobMessage}
\title{JobMessage Object}
\usage{
JobMessage(id = NULL, time = NULL, messageText = NULL,
  messageImportance = NULL)
}
\arguments{
\item{id}{Identifies the message}

\item{time}{The timestamp of the message}

\item{messageText}{The text of the message}

\item{messageImportance}{Importance level of the message}
}
\value{
JobMessage object
}
\description{
JobMessage Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A particular message pertaining to a Dataflow job.
}

