% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{JobMetrics}
\alias{JobMetrics}
\title{JobMetrics Object}
\usage{
JobMetrics(metricTime = NULL, metrics = NULL)
}
\arguments{
\item{metricTime}{Timestamp as of which metric values are current}

\item{metrics}{All metrics for this job}
}
\value{
JobMetrics object
}
\description{
JobMetrics Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JobMetrics contains a collection of metrics descibing the detailed progressof a Dataflow job. Metrics correspond to user-defined and system-definedmetrics in the job.This resource captures only the most recent values of each metric;time-series data can be queried for them (under the same metric names)from Cloud Monitoring.
}
