% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{KeyRangeLocation}
\alias{KeyRangeLocation}
\title{KeyRangeLocation Object}
\usage{
KeyRangeLocation(start = NULL, dataDisk = NULL,
  persistentDirectory = NULL, end = NULL, deliveryEndpoint = NULL)
}
\arguments{
\item{start}{The start (inclusive) of the key range}

\item{dataDisk}{The name of the data disk where data for this range is stored}

\item{persistentDirectory}{The location of the persistent state for this range, as a}

\item{end}{The end (exclusive) of the key range}

\item{deliveryEndpoint}{The physical location of this range assignment to be used for}
}
\value{
KeyRangeLocation object
}
\description{
KeyRangeLocation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Location information for a specific key-range of a sharded computation.Currently we only support UTF-8 character splits to simplify encoding intoJSON.
}
