% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{LeaseWorkItemRequest}
\alias{LeaseWorkItemRequest}
\title{LeaseWorkItemRequest Object}
\usage{
LeaseWorkItemRequest(workItemTypes = NULL, workerCapabilities = NULL,
  requestedLeaseDuration = NULL, currentWorkerTime = NULL,
  workerId = NULL)
}
\arguments{
\item{workItemTypes}{Filter for WorkItem type}

\item{workerCapabilities}{Worker capabilities}

\item{requestedLeaseDuration}{The initial lease period}

\item{currentWorkerTime}{The current timestamp at the worker}

\item{workerId}{Identifies the worker leasing work -- typically the ID of the virtual machine running the worker}
}
\value{
LeaseWorkItemRequest object
}
\description{
LeaseWorkItemRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request to lease WorkItems.
}

