% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{LeaseWorkItemResponse}
\alias{LeaseWorkItemResponse}
\title{LeaseWorkItemResponse Object}
\usage{
LeaseWorkItemResponse(workItems = NULL)
}
\arguments{
\item{workItems}{A list of the leased WorkItems}
}
\value{
LeaseWorkItemResponse object
}
\description{
LeaseWorkItemResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response to a request to lease WorkItems.
}
