% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{ListJobMessagesResponse}
\alias{ListJobMessagesResponse}
\title{ListJobMessagesResponse Object}
\usage{
ListJobMessagesResponse(jobMessages = NULL, nextPageToken = NULL)
}
\arguments{
\item{jobMessages}{Messages in ascending timestamp order}

\item{nextPageToken}{The token to obtain the next page of results if there are more}
}
\value{
ListJobMessagesResponse object
}
\description{
ListJobMessagesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response to a request to list job messages.
}
