% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{ListJobsResponse}
\alias{ListJobsResponse}
\title{ListJobsResponse Object}
\usage{
ListJobsResponse(jobs = NULL, nextPageToken = NULL)
}
\arguments{
\item{jobs}{A subset of the requested job information}

\item{nextPageToken}{Set if there may be more results than fit in this response}
}
\value{
ListJobsResponse object
}
\description{
ListJobsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response to a request to list Dataflow jobs. This may be a partial response, depending on the page size in the ListJobsRequest.
}

