% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{MapTask}
\alias{MapTask}
\title{MapTask Object}
\usage{
MapTask(instructions = NULL, systemName = NULL, stageName = NULL)
}
\arguments{
\item{instructions}{The instructions in the MapTask}

\item{systemName}{System-defined name of this MapTask}

\item{stageName}{System-defined name of the stage containing this MapTask}
}
\value{
MapTask object
}
\description{
MapTask Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
MapTask consists of an ordered set of instructions, each of which describes one particular low-level operation for the worker to perform in order to accomplish the MapTask's WorkItem. Each instruction must appear in the list before any instructions which depends on its output.
}

