% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{MetricShortId}
\alias{MetricShortId}
\title{MetricShortId Object}
\usage{
MetricShortId(metricIndex = NULL, shortId = NULL)
}
\arguments{
\item{metricIndex}{The index of the corresponding metric in the ReportWorkItemStatusRequest}

\item{shortId}{The service-generated short identifier for the metric}
}
\value{
MetricShortId object
}
\description{
MetricShortId Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The metric short id is returned to the user alongside an offset into ReportWorkItemStatusRequest
}

