% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{MetricStructuredName.context}
\alias{MetricStructuredName.context}
\title{MetricStructuredName.context Object}
\usage{
MetricStructuredName.context()
}
\value{
MetricStructuredName.context object
}
\description{
MetricStructuredName.context Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Zero or more labeled fields which identify the part of the job thismetric is associated with, such as the name of a step or collection.For example, built-in counters associated with steps will havecontext['step'] = <step-name>. Counters associated with PCollectionsin the SDK will have context['pcollection'] = <pcollection-name>.
}
\seealso{
Other MetricStructuredName functions: \code{\link{MetricStructuredName}}
}
