% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{MetricStructuredName}
\alias{MetricStructuredName}
\title{MetricStructuredName Object}
\usage{
MetricStructuredName(MetricStructuredName.context = NULL, origin = NULL,
  name = NULL, context = NULL)
}
\arguments{
\item{MetricStructuredName.context}{The \link{MetricStructuredName.context} object or list of objects}

\item{origin}{Origin (namespace) of metric name}

\item{name}{Worker-defined metric name}

\item{context}{Zero or more labeled fields which identify the part of the job this}
}
\value{
MetricStructuredName object
}
\description{
MetricStructuredName Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Identifies a metric, by describing the source which generated themetric.
}
\seealso{
Other MetricStructuredName functions: \code{\link{MetricStructuredName.context}}
}
