% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{MetricUpdate}
\alias{MetricUpdate}
\title{MetricUpdate Object}
\usage{
MetricUpdate(name = NULL, cumulative = NULL, scalar = NULL,
  meanSum = NULL, meanCount = NULL, set = NULL, internal = NULL,
  updateTime = NULL)
}
\arguments{
\item{name}{Name of the metric}

\item{cumulative}{True if this metric is reported as the total cumulative aggregate value accumulated since the worker started working on this WorkItem}

\item{scalar}{Worker-computed aggregate value for aggregation kinds 'Sum', 'Max', 'Min', 'And', and 'Or'}

\item{meanSum}{Worker-computed aggregate value for the 'Mean' aggregation kind}

\item{meanCount}{Worker-computed aggregate value for the 'Mean' aggregation kind}

\item{set}{Worker-computed aggregate value for the 'Set' aggregation kind}

\item{internal}{Worker-computed aggregate value for internal use by the Dataflow service}

\item{updateTime}{Timestamp associated with the metric value}
}
\value{
MetricUpdate object
}
\description{
MetricUpdate Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Describes the state of a metric.
}

