% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{MountedDataDisk}
\alias{MountedDataDisk}
\title{MountedDataDisk Object}
\usage{
MountedDataDisk(dataDisk = NULL)
}
\arguments{
\item{dataDisk}{The name of the data disk}
}
\value{
MountedDataDisk object
}
\description{
MountedDataDisk Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Describes mounted data disk.
}
