% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{MultiOutputInfo}
\alias{MultiOutputInfo}
\title{MultiOutputInfo Object}
\usage{
MultiOutputInfo(tag = NULL)
}
\arguments{
\item{tag}{The id of the tag the user code will emit to this output by; this}
}
\value{
MultiOutputInfo object
}
\description{
MultiOutputInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Information about an output of a multi-output DoFn.
}
