% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{NameAndKind}
\alias{NameAndKind}
\title{NameAndKind Object}
\usage{
NameAndKind(name = NULL)
}
\arguments{
\item{name}{Name of the counter}
}
\value{
NameAndKind object
}
\description{
NameAndKind Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Basic metadata about a counter.
}

