% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{Package}
\alias{Package}
\title{Package Object}
\usage{
Package(location = NULL, name = NULL)
}
\arguments{
\item{location}{The resource to read the package from}

\item{name}{The name of the package}
}
\value{
Package object
}
\description{
Package Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The packages that must be installed in order for a worker to run thesteps of the Cloud Dataflow job that will be assigned to its workerpool.This is the mechanism by which the Cloud Dataflow SDK causes code tobe loaded onto the workers. For example, the Cloud Dataflow Java SDKmight use this to install jars containing the user's code and all of thevarious dependencies (libraries, data files, etc.) required in orderfor that code to run.
}
