% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{ParallelInstruction}
\alias{ParallelInstruction}
\title{ParallelInstruction Object}
\usage{
ParallelInstruction(outputs = NULL, name = NULL, parDo = NULL,
  read = NULL, flatten = NULL, originalName = NULL, write = NULL,
  systemName = NULL, partialGroupByKey = NULL)
}
\arguments{
\item{outputs}{Describes the outputs of the instruction}

\item{name}{User-provided name of this operation}

\item{parDo}{Additional information for ParDo instructions}

\item{read}{Additional information for Read instructions}

\item{flatten}{Additional information for Flatten instructions}

\item{originalName}{System-defined name for the operation in the original workflow graph}

\item{write}{Additional information for Write instructions}

\item{systemName}{System-defined name of this operation}

\item{partialGroupByKey}{Additional information for PartialGroupByKey instructions}
}
\value{
ParallelInstruction object
}
\description{
ParallelInstruction Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Describes a particular operation comprising a MapTask.
}
