% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{ParDoInstruction}
\alias{ParDoInstruction}
\title{ParDoInstruction Object}
\usage{
ParDoInstruction(ParDoInstruction.userFn = NULL, input = NULL,
  sideInputs = NULL, userFn = NULL, numOutputs = NULL,
  multiOutputInfos = NULL)
}
\arguments{
\item{ParDoInstruction.userFn}{The \link{ParDoInstruction.userFn} object or list of objects}

\item{input}{The input}

\item{sideInputs}{Zero or more side inputs}

\item{userFn}{The user function to invoke}

\item{numOutputs}{The number of outputs}

\item{multiOutputInfos}{Information about each of the outputs, if user_fn is a MultiDoFn}
}
\value{
ParDoInstruction object
}
\description{
ParDoInstruction Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An instruction that does a ParDo operation. Takes one main input and zero or more side inputs, and produces zero or more outputs. Runs user code.
}
\seealso{
Other ParDoInstruction functions: \code{\link{ParDoInstruction.userFn}}
}

