% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{PartialGroupByKeyInstruction.inputElementCodec}
\alias{PartialGroupByKeyInstruction.inputElementCodec}
\title{PartialGroupByKeyInstruction.inputElementCodec Object}
\usage{
PartialGroupByKeyInstruction.inputElementCodec()
}
\value{
PartialGroupByKeyInstruction.inputElementCodec object
}
\description{
PartialGroupByKeyInstruction.inputElementCodec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The codec to use for interpreting an element in the input PTable.
}
\seealso{
Other PartialGroupByKeyInstruction functions: \code{\link{PartialGroupByKeyInstruction.valueCombiningFn}},
  \code{\link{PartialGroupByKeyInstruction}}
}

