% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{PartialGroupByKeyInstruction}
\alias{PartialGroupByKeyInstruction}
\title{PartialGroupByKeyInstruction Object}
\usage{

  PartialGroupByKeyInstruction(PartialGroupByKeyInstruction.inputElementCodec = NULL,
  PartialGroupByKeyInstruction.valueCombiningFn = NULL, input = NULL,
  inputElementCodec = NULL, valueCombiningFn = NULL, sideInputs = NULL,
  originalCombineValuesStepName = NULL,
  originalCombineValuesInputStoreName = NULL)
}
\arguments{
\item{PartialGroupByKeyInstruction.inputElementCodec}{The \link{PartialGroupByKeyInstruction.inputElementCodec} object or list of objects}

\item{PartialGroupByKeyInstruction.valueCombiningFn}{The \link{PartialGroupByKeyInstruction.valueCombiningFn} object or list of objects}

\item{input}{Describes the input to the partial group-by-key instruction}

\item{inputElementCodec}{The codec to use for interpreting an element in the input PTable}

\item{valueCombiningFn}{The value combining function to invoke}

\item{sideInputs}{Zero or more side inputs}

\item{originalCombineValuesStepName}{If this instruction includes a combining function, this is the name of the CombineValues instruction lifted into this instruction}

\item{originalCombineValuesInputStoreName}{If this instruction includes a combining function this is the name of the intermediate store between the GBK and the CombineValues}
}
\value{
PartialGroupByKeyInstruction object
}
\description{
PartialGroupByKeyInstruction Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An instruction that does a partial group-by-key. One input and one output.
}
\seealso{
Other PartialGroupByKeyInstruction functions: \code{\link{PartialGroupByKeyInstruction.inputElementCodec}},
  \code{\link{PartialGroupByKeyInstruction.valueCombiningFn}}
}

