% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{PipelineDescription}
\alias{PipelineDescription}
\title{PipelineDescription Object}
\usage{
PipelineDescription(displayData = NULL, executionPipelineStage = NULL,
  originalPipelineTransform = NULL)
}
\arguments{
\item{displayData}{Pipeline level display data}

\item{executionPipelineStage}{Description of each stage of execution of the pipeline}

\item{originalPipelineTransform}{Description of each transform in the pipeline and collections between them}
}
\value{
PipelineDescription object
}
\description{
PipelineDescription Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A descriptive representation of submitted pipeline as well as the executedform.  This data is provided by the Dataflow service for ease of visualizingthe pipeline and interpretting Dataflow provided metrics.
}
