% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_functions.R
\name{projects.workerMessages}
\alias{projects.workerMessages}
\title{Send a worker_message to the service.}
\usage{
projects.workerMessages(SendWorkerMessagesRequest, projectId)
}
\arguments{
\item{SendWorkerMessagesRequest}{The \link{SendWorkerMessagesRequest} object to pass to this method}

\item{projectId}{The project to send the WorkerMessages to}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/userinfo.email
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/userinfo.email)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/dataflow}{Google Documentation}

Other SendWorkerMessagesRequest functions: \code{\link{SendWorkerMessagesRequest}}
}
