% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{PubsubLocation}
\alias{PubsubLocation}
\title{PubsubLocation Object}
\usage{
PubsubLocation(withAttributes = NULL, idLabel = NULL,
  timestampLabel = NULL, topic = NULL, subscription = NULL,
  dropLateData = NULL, trackingSubscription = NULL)
}
\arguments{
\item{withAttributes}{If true, then the client has requested to get pubsub attributes}

\item{idLabel}{If set, contains a pubsub label from which to extract record ids}

\item{timestampLabel}{If set, contains a pubsub label from which to extract record timestamps}

\item{topic}{A pubsub topic, in the form of}

\item{subscription}{A pubsub subscription, in the form of}

\item{dropLateData}{Indicates whether the pipeline allows late-arriving data}

\item{trackingSubscription}{If set, specifies the pubsub subscription that will be used for tracking}
}
\value{
PubsubLocation object
}
\description{
PubsubLocation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Identifies a pubsub location to use for transferring data into orout of a streaming Dataflow job.
}
