% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{ReadInstruction}
\alias{ReadInstruction}
\title{ReadInstruction Object}
\usage{
ReadInstruction(source = NULL)
}
\arguments{
\item{source}{The source to read from}
}
\value{
ReadInstruction object
}
\description{
ReadInstruction Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An instruction that reads records. Takes no inputs, produces one output.
}

