% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{ReportedParallelism}
\alias{ReportedParallelism}
\title{ReportedParallelism Object}
\usage{
ReportedParallelism(isInfinite = NULL, value = NULL)
}
\arguments{
\item{isInfinite}{Specifies whether the parallelism is infinite}

\item{value}{Specifies the level of parallelism in case it is finite}
}
\value{
ReportedParallelism object
}
\description{
ReportedParallelism Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents the level of parallelism in a WorkItem's input, reported by the worker.
}

