% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{ReportWorkItemStatusRequest}
\alias{ReportWorkItemStatusRequest}
\title{ReportWorkItemStatusRequest Object}
\usage{
ReportWorkItemStatusRequest(location = NULL, workItemStatuses = NULL,
  workerId = NULL, currentWorkerTime = NULL)
}
\arguments{
\item{location}{The location which contains the WorkItem's job}

\item{workItemStatuses}{The order is unimportant, except that the order of the}

\item{workerId}{The ID of the worker reporting the WorkItem status}

\item{currentWorkerTime}{The current timestamp at the worker}
}
\value{
ReportWorkItemStatusRequest object
}
\description{
ReportWorkItemStatusRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request to report the status of WorkItems.
}
