% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{ReportWorkItemStatusResponse}
\alias{ReportWorkItemStatusResponse}
\title{ReportWorkItemStatusResponse Object}
\usage{
ReportWorkItemStatusResponse(workItemServiceStates = NULL)
}
\arguments{
\item{workItemServiceStates}{A set of messages indicating the service-side state for each}
}
\value{
ReportWorkItemStatusResponse object
}
\description{
ReportWorkItemStatusResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response from a request to report the status of WorkItems.
}
