% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{ResourceUtilizationReport}
\alias{ResourceUtilizationReport}
\title{ResourceUtilizationReport Object}
\usage{
ResourceUtilizationReport(cpuTime = NULL)
}
\arguments{
\item{cpuTime}{CPU utilization samples}
}
\value{
ResourceUtilizationReport object
}
\description{
ResourceUtilizationReport Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Worker metrics exported from workers. This contains resource utilizationmetrics accumulated from a variety of sources. For more information, seego/df-resource-signals.
}
