% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{ResourceUtilizationReportResponse}
\alias{ResourceUtilizationReportResponse}
\title{ResourceUtilizationReportResponse Object}
\usage{
ResourceUtilizationReportResponse()
}
\value{
ResourceUtilizationReportResponse object
}
\description{
ResourceUtilizationReportResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Service-side response to WorkerMessage reporting resource utilization.
}
