% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{RuntimeEnvironment}
\alias{RuntimeEnvironment}
\title{Google Dataflow API Objects 
Manages Google Cloud Dataflow projects on Google Cloud Platform.}
\usage{
RuntimeEnvironment(maxWorkers = NULL, zone = NULL,
  serviceAccountEmail = NULL, tempLocation = NULL,
  bypassTempDirValidation = NULL)
}
\arguments{
\item{maxWorkers}{The maximum number of Google Compute Engine instances to be made}

\item{zone}{The Compute Engine [availability zone](https://cloud}

\item{serviceAccountEmail}{The email address of the service account to run the job as}

\item{tempLocation}{The Cloud Storage path to use for temporary files}

\item{bypassTempDirValidation}{Whether to bypass the safety checks for the job's temporary directory}
}
\value{
RuntimeEnvironment object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 19:41:42
filename: /Users/mark/dev/R/autoGoogleAPI/googledataflowv1b3.auto/R/dataflow_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
RuntimeEnvironment Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The environment values to set at runtime.
}
