% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{SendDebugCaptureRequest}
\alias{SendDebugCaptureRequest}
\title{SendDebugCaptureRequest Object}
\usage{
SendDebugCaptureRequest(data = NULL, componentId = NULL, workerId = NULL)
}
\arguments{
\item{data}{The encoded debug information}

\item{componentId}{The internal component id for which debug information is sent}

\item{workerId}{The worker id, i}
}
\value{
SendDebugCaptureRequest object
}
\description{
SendDebugCaptureRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request to send encoded debug information.
}
