% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{SendDebugCaptureResponse}
\alias{SendDebugCaptureResponse}
\title{SendDebugCaptureResponse Object}
\usage{
SendDebugCaptureResponse()
}
\value{
SendDebugCaptureResponse object
}
\description{
SendDebugCaptureResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response to a send capture request. nothing
}

