% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{SendWorkerMessagesRequest}
\alias{SendWorkerMessagesRequest}
\title{SendWorkerMessagesRequest Object}
\usage{
SendWorkerMessagesRequest(workerMessages = NULL)
}
\arguments{
\item{workerMessages}{The WorkerMessages to send}
}
\value{
SendWorkerMessagesRequest object
}
\description{
SendWorkerMessagesRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A request for sending worker messages to the service.
}
\seealso{
Other SendWorkerMessagesRequest functions: \code{\link{projects.workerMessages}}
}

