% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{SendWorkerMessagesResponse}
\alias{SendWorkerMessagesResponse}
\title{SendWorkerMessagesResponse Object}
\usage{
SendWorkerMessagesResponse(workerMessageResponses = NULL)
}
\arguments{
\item{workerMessageResponses}{The servers response to the worker messages}
}
\value{
SendWorkerMessagesResponse object
}
\description{
SendWorkerMessagesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response to the worker messages.
}
