% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{SeqMapTask}
\alias{SeqMapTask}
\title{SeqMapTask Object}
\usage{
SeqMapTask(SeqMapTask.userFn = NULL, inputs = NULL, userFn = NULL,
  outputInfos = NULL, name = NULL, systemName = NULL, stageName = NULL)
}
\arguments{
\item{SeqMapTask.userFn}{The \link{SeqMapTask.userFn} object or list of objects}

\item{inputs}{Information about each of the inputs}

\item{userFn}{The user function to invoke}

\item{outputInfos}{Information about each of the outputs}

\item{name}{The user-provided name of the SeqDo operation}

\item{systemName}{System-defined name of the SeqDo operation}

\item{stageName}{System-defined name of the stage containing the SeqDo operation}
}
\value{
SeqMapTask object
}
\description{
SeqMapTask Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Describes a particular function to invoke.
}
\seealso{
Other SeqMapTask functions: \code{\link{SeqMapTask.userFn}}
}

