% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{SeqMapTask.userFn}
\alias{SeqMapTask.userFn}
\title{SeqMapTask.userFn Object}
\usage{
SeqMapTask.userFn()
}
\value{
SeqMapTask.userFn object
}
\description{
SeqMapTask.userFn Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The user function to invoke.
}
\seealso{
Other SeqMapTask functions: \code{\link{SeqMapTask}}
}
