% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{ShellTask}
\alias{ShellTask}
\title{ShellTask Object}
\usage{
ShellTask(command = NULL, exitCode = NULL)
}
\arguments{
\item{command}{The shell command to run}

\item{exitCode}{Exit code for the task}
}
\value{
ShellTask object
}
\description{
ShellTask Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A task which consists of a shell command for the worker to execute.
}

