% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{SideInputInfo.kind}
\alias{SideInputInfo.kind}
\title{SideInputInfo.kind Object}
\usage{
SideInputInfo.kind()
}
\value{
SideInputInfo.kind object
}
\description{
SideInputInfo.kind Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
How to interpret the source element(s) as a side input value.
}
\seealso{
Other SideInputInfo functions: \code{\link{SideInputInfo}}
}

