% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{SideInputInfo}
\alias{SideInputInfo}
\title{SideInputInfo Object}
\usage{
SideInputInfo(SideInputInfo.kind = NULL, sources = NULL, tag = NULL)
}
\arguments{
\item{SideInputInfo.kind}{The \link{SideInputInfo.kind} object or list of objects}

\item{sources}{The source(s) to read element(s) from to get the value of this side input}

\item{tag}{The id of the tag the user code will access this side input by;}
}
\value{
SideInputInfo object
}
\description{
SideInputInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Information about a side input of a DoFn or an input of a SeqDoFn.
}
\seealso{
Other SideInputInfo functions: \code{\link{SideInputInfo.kind}}
}
