% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{Sink.codec}
\alias{Sink.codec}
\title{Sink.codec Object}
\usage{
Sink.codec()
}
\value{
Sink.codec object
}
\description{
Sink.codec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The codec to use to encode data written to the sink.
}
\seealso{
Other Sink functions: \code{\link{Sink.spec}},
  \code{\link{Sink}}
}
