% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{Sink}
\alias{Sink}
\title{Sink Object}
\usage{
Sink(Sink.codec = NULL, Sink.spec = NULL, codec = NULL, spec = NULL)
}
\arguments{
\item{Sink.codec}{The \link{Sink.codec} object or list of objects}

\item{Sink.spec}{The \link{Sink.spec} object or list of objects}

\item{codec}{The codec to use to encode data written to the sink}

\item{spec}{The sink to write to, plus its parameters}
}
\value{
Sink object
}
\description{
Sink Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A sink that records can be encoded and written to.
}
\seealso{
Other Sink functions: \code{\link{Sink.codec}},
  \code{\link{Sink.spec}}
}
