% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{Sink.spec}
\alias{Sink.spec}
\title{Sink.spec Object}
\usage{
Sink.spec()
}
\value{
Sink.spec object
}
\description{
Sink.spec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The sink to write to, plus its parameters.
}
\seealso{
Other Sink functions: \code{\link{Sink.codec}},
  \code{\link{Sink}}
}
