% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{Source.codec}
\alias{Source.codec}
\title{Source.codec Object}
\usage{
Source.codec()
}
\value{
Source.codec object
}
\description{
Source.codec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The codec to use to decode data read from the source.
}
\seealso{
Other Source functions: \code{\link{Source.spec}},
  \code{\link{Source}}
}
