% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{Source.spec}
\alias{Source.spec}
\title{Source.spec Object}
\usage{
Source.spec()
}
\value{
Source.spec object
}
\description{
Source.spec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The source to read from, plus its parameters.
}
\seealso{
Other Source functions: \code{\link{Source.codec}},
  \code{\link{Source}}
}

