% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{SourceGetMetadataRequest}
\alias{SourceGetMetadataRequest}
\title{SourceGetMetadataRequest Object}
\usage{
SourceGetMetadataRequest(source = NULL)
}
\arguments{
\item{source}{Specification of the source whose metadata should be computed}
}
\value{
SourceGetMetadataRequest object
}
\description{
SourceGetMetadataRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A request to compute the SourceMetadata of a Source.
}
