% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{SourceGetMetadataResponse}
\alias{SourceGetMetadataResponse}
\title{SourceGetMetadataResponse Object}
\usage{
SourceGetMetadataResponse(metadata = NULL)
}
\arguments{
\item{metadata}{The computed metadata}
}
\value{
SourceGetMetadataResponse object
}
\description{
SourceGetMetadataResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The result of a SourceGetMetadataOperation.
}
