% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{SourceOperationRequest}
\alias{SourceOperationRequest}
\title{SourceOperationRequest Object}
\usage{
SourceOperationRequest(getMetadata = NULL, split = NULL)
}
\arguments{
\item{getMetadata}{Information about a request to get metadata about a source}

\item{split}{Information about a request to split a source}
}
\value{
SourceOperationRequest object
}
\description{
SourceOperationRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A work item that represents the different operations that can beperformed on a user-defined Source specification.
}
