% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{SourceOperationResponse}
\alias{SourceOperationResponse}
\title{SourceOperationResponse Object}
\usage{
SourceOperationResponse(split = NULL, getMetadata = NULL)
}
\arguments{
\item{split}{A response to a request to split a source}

\item{getMetadata}{A response to a request to get metadata about a source}
}
\value{
SourceOperationResponse object
}
\description{
SourceOperationResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The result of a SourceOperationRequest, specified inReportWorkItemStatusRequest.source_operation when the work itemis completed.
}
