% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{SourceSplitOptions}
\alias{SourceSplitOptions}
\title{SourceSplitOptions Object}
\usage{
SourceSplitOptions(desiredBundleSizeBytes = NULL,
  desiredShardSizeBytes = NULL)
}
\arguments{
\item{desiredBundleSizeBytes}{The source should be split into a set of bundles where the estimated size of each is approximately this many bytes}

\item{desiredShardSizeBytes}{DEPRECATED in favor of desired_bundle_size_bytes}
}
\value{
SourceSplitOptions object
}
\description{
SourceSplitOptions Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Hints for splitting a Source into bundles (parts for parallel processing) using SourceSplitRequest.
}

