% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{SourceSplitRequest}
\alias{SourceSplitRequest}
\title{SourceSplitRequest Object}
\usage{
SourceSplitRequest(source = NULL, options = NULL)
}
\arguments{
\item{source}{Specification of the source to be split}

\item{options}{Hints for tuning the splitting process}
}
\value{
SourceSplitRequest object
}
\description{
SourceSplitRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents the operation to split a high-level Source specificationinto bundles (parts for parallel processing).At a high level, splitting of a source into bundles happens as follows:SourceSplitRequest is applied to the source. If it returnsSOURCE_SPLIT_OUTCOME_USE_CURRENT, no further splitting happens and the sourceis used 'as is'. Otherwise, splitting is applied recursively to eachproduced DerivedSource.As an optimization, for any Source, if its does_not_need_splitting istrue, the framework assumes that splitting this source would returnSOURCE_SPLIT_OUTCOME_USE_CURRENT, and doesn't initiate a SourceSplitRequest.This applies both to the initial source being split and to bundlesproduced from it.
}
