% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{SourceSplitResponse}
\alias{SourceSplitResponse}
\title{SourceSplitResponse Object}
\usage{
SourceSplitResponse(outcome = NULL, bundles = NULL, shards = NULL)
}
\arguments{
\item{outcome}{Indicates whether splitting happened and produced a list of bundles}

\item{bundles}{If outcome is SPLITTING_HAPPENED, then this is a list of bundles into which the source was split}

\item{shards}{DEPRECATED in favor of bundles}
}
\value{
SourceSplitResponse object
}
\description{
SourceSplitResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response to a SourceSplitRequest.
}

