% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{SplitInt64}
\alias{SplitInt64}
\title{SplitInt64 Object}
\usage{
SplitInt64(lowBits = NULL, highBits = NULL)
}
\arguments{
\item{lowBits}{The low order bits: n & 0xffffffff}

\item{highBits}{The high order bits, including the sign: n >> 32}
}
\value{
SplitInt64 object
}
\description{
SplitInt64 Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A representation of an int64, n, that is immune to precision loss when encoded in JSON.
}

