% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{StageSource}
\alias{StageSource}
\title{StageSource Object}
\usage{
StageSource(name = NULL, originalUserTransformOrCollection = NULL,
  sizeBytes = NULL, userName = NULL)
}
\arguments{
\item{name}{Dataflow service generated name for this source}

\item{originalUserTransformOrCollection}{User name for the original user transform or collection with which this}

\item{sizeBytes}{Size of the source, if measurable}

\item{userName}{Human-readable name for this source; may be user or system generated}
}
\value{
StageSource object
}
\description{
StageSource Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Description of an input or output of an execution stage.
}
