% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{StateFamilyConfig}
\alias{StateFamilyConfig}
\title{StateFamilyConfig Object}
\usage{
StateFamilyConfig(stateFamily = NULL, isRead = NULL)
}
\arguments{
\item{stateFamily}{The state family value}

\item{isRead}{If true, this family corresponds to a read operation}
}
\value{
StateFamilyConfig object
}
\description{
StateFamilyConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
State family configuration.
}

