% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{Step.properties}
\alias{Step.properties}
\title{Step.properties Object}
\usage{
Step.properties()
}
\value{
Step.properties object
}
\description{
Step.properties Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Named properties associated with the step. Each kind of predefined step has its own required set of properties.
}
\seealso{
Other Step functions: \code{\link{Step}}
}

