% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{Step}
\alias{Step}
\title{Step Object}
\usage{
Step(Step.properties = NULL, properties = NULL, name = NULL)
}
\arguments{
\item{Step.properties}{The \link{Step.properties} object or list of objects}

\item{properties}{Named properties associated with the step}

\item{name}{The name that identifies the step}
}
\value{
Step object
}
\description{
Step Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Defines a particular step within a Cloud Dataflow job.A job consists of multiple steps, each of which performs somespecific operation as part of the overall job.  Data is typicallypassed from one step to another as part of the job.Here's an example of a sequence of steps which together implement aMap-Reduce job:  * Read a collection of data from some source, parsing the    collection's elements.  * Validate the elements.  * Apply a user-defined function to map each element to some value    and extract an element-specific key value.  * Group elements with the same key into a single element with    that key, transforming a multiply-keyed collection into a    uniquely-keyed collection.  * Write the elements out to some data sink.Note that the Cloud Dataflow service may be used to run many differenttypes of jobs, not just Map-Reduce.
}
\seealso{
Other Step functions: \code{\link{Step.properties}}
}
