% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{StreamingComputationConfig}
\alias{StreamingComputationConfig}
\title{StreamingComputationConfig Object}
\usage{
StreamingComputationConfig(computationId = NULL, systemName = NULL,
  stageName = NULL, instructions = NULL)
}
\arguments{
\item{computationId}{Unique identifier for this computation}

\item{systemName}{System defined name for this computation}

\item{stageName}{Stage name of this computation}

\item{instructions}{Instructions that comprise the computation}
}
\value{
StreamingComputationConfig object
}
\description{
StreamingComputationConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Configuration information for a single streaming computation.
}

