% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{StreamingComputationRanges}
\alias{StreamingComputationRanges}
\title{StreamingComputationRanges Object}
\usage{
StreamingComputationRanges(rangeAssignments = NULL, computationId = NULL)
}
\arguments{
\item{rangeAssignments}{Data disk assignments for ranges from this computation}

\item{computationId}{The ID of the computation}
}
\value{
StreamingComputationRanges object
}
\description{
StreamingComputationRanges Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Describes full or partial data disk assignment information of the computationranges.
}
