% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{StreamingComputationTask}
\alias{StreamingComputationTask}
\title{StreamingComputationTask Object}
\usage{
StreamingComputationTask(taskType = NULL, dataDisks = NULL,
  computationRanges = NULL)
}
\arguments{
\item{taskType}{A type of streaming computation task}

\item{dataDisks}{Describes the set of data disks this task should apply to}

\item{computationRanges}{Contains ranges of a streaming computation this task should apply to}
}
\value{
StreamingComputationTask object
}
\description{
StreamingComputationTask Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A task which describes what action should be performed for the specified streaming computation ranges.
}

