% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{StreamingConfigTask}
\alias{StreamingConfigTask}
\title{StreamingConfigTask Object}
\usage{
StreamingConfigTask(StreamingConfigTask.userStepToStateFamilyNameMap = NULL,
  streamingComputationConfigs = NULL, userStepToStateFamilyNameMap = NULL)
}
\arguments{
\item{StreamingConfigTask.userStepToStateFamilyNameMap}{The \link{StreamingConfigTask.userStepToStateFamilyNameMap} object or list of objects}

\item{streamingComputationConfigs}{Set of computation configuration information}

\item{userStepToStateFamilyNameMap}{Map from user step names to state families}
}
\value{
StreamingConfigTask object
}
\description{
StreamingConfigTask Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A task that carries configuration information for streaming computations.
}
\seealso{
Other StreamingConfigTask functions: \code{\link{StreamingConfigTask.userStepToStateFamilyNameMap}}
}

