% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{StreamingConfigTask.userStepToStateFamilyNameMap}
\alias{StreamingConfigTask.userStepToStateFamilyNameMap}
\title{StreamingConfigTask.userStepToStateFamilyNameMap Object}
\usage{
StreamingConfigTask.userStepToStateFamilyNameMap()
}
\value{
StreamingConfigTask.userStepToStateFamilyNameMap object
}
\description{
StreamingConfigTask.userStepToStateFamilyNameMap Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Map from user step names to state families.
}
\seealso{
Other StreamingConfigTask functions: \code{\link{StreamingConfigTask}}
}

