% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{StreamingSetupTask}
\alias{StreamingSetupTask}
\title{StreamingSetupTask Object}
\usage{
StreamingSetupTask(receiveWorkPort = NULL, workerHarnessPort = NULL,
  streamingComputationTopology = NULL, drain = NULL)
}
\arguments{
\item{receiveWorkPort}{The TCP port on which the worker should listen for messages from other streaming computation workers}

\item{workerHarnessPort}{The TCP port used by the worker to communicate with the Dataflow worker harness}

\item{streamingComputationTopology}{The global topology of the streaming Dataflow job}

\item{drain}{The user has requested drain}
}
\value{
StreamingSetupTask object
}
\description{
StreamingSetupTask Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A task which initializes part of a streaming Dataflow job.
}

