% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{StreamingStageLocation}
\alias{StreamingStageLocation}
\title{StreamingStageLocation Object}
\usage{
StreamingStageLocation(streamId = NULL)
}
\arguments{
\item{streamId}{Identifies the particular stream within the streaming Dataflow job}
}
\value{
StreamingStageLocation object
}
\description{
StreamingStageLocation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Identifies the location of a streaming computation stage, for stage-to-stage communication.
}

