% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{StreamLocation}
\alias{StreamLocation}
\title{StreamLocation Object}
\usage{
StreamLocation(customSourceLocation = NULL, streamingStageLocation = NULL,
  pubsubLocation = NULL, sideInputLocation = NULL)
}
\arguments{
\item{customSourceLocation}{The stream is a custom source}

\item{streamingStageLocation}{The stream is part of another computation within the current}

\item{pubsubLocation}{The stream is a pubsub stream}

\item{sideInputLocation}{The stream is a streaming side input}
}
\value{
StreamLocation object
}
\description{
StreamLocation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Describes a stream of data, either as input to be processed or asoutput of a streaming Dataflow job.
}
