% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{StringList}
\alias{StringList}
\title{StringList Object}
\usage{
StringList(elements = NULL)
}
\arguments{
\item{elements}{Elements of the list}
}
\value{
StringList object
}
\description{
StringList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A metric value representing a list of strings.
}

