% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{TaskRunnerSettings}
\alias{TaskRunnerSettings}
\title{TaskRunnerSettings Object}
\usage{
TaskRunnerSettings(taskUser = NULL, taskGroup = NULL, oauthScopes = NULL,
  baseUrl = NULL, dataflowApiVersion = NULL,
  parallelWorkerSettings = NULL, baseTaskDir = NULL,
  continueOnException = NULL, logToSerialconsole = NULL,
  alsologtostderr = NULL, logUploadLocation = NULL, logDir = NULL,
  tempStoragePrefix = NULL, harnessCommand = NULL,
  workflowFileName = NULL, commandlinesFileName = NULL, vmId = NULL,
  languageHint = NULL, streamingWorkerMainClass = NULL)
}
\arguments{
\item{taskUser}{The UNIX user ID on the worker VM to use for tasks launched by taskrunner; e}

\item{taskGroup}{The UNIX group ID on the worker VM to use for tasks launched by taskrunner; e}

\item{oauthScopes}{OAuth2 scopes to be requested by the taskrunner in order to access the dataflow API}

\item{baseUrl}{The base URL for the taskrunner to use when accessing Google Cloud APIs}

\item{dataflowApiVersion}{API version of endpoint, e}

\item{parallelWorkerSettings}{Settings to pass to the parallel worker harness}

\item{baseTaskDir}{Location on the worker for task-specific subdirectories}

\item{continueOnException}{Do we continue taskrunner if an exception is hit?}

\item{logToSerialconsole}{Send taskrunner log into to Google Compute Engine VM serial console?}

\item{alsologtostderr}{Also send taskrunner log info to stderr?}

\item{logUploadLocation}{Indicates where to put logs}

\item{logDir}{Directory on the VM to store logs}

\item{tempStoragePrefix}{The prefix of the resources the taskrunner should use for temporary storage}

\item{harnessCommand}{Command to launch the worker harness}

\item{workflowFileName}{Store the workflow in this file}

\item{commandlinesFileName}{Store preprocessing commands in this file}

\item{vmId}{ID string of VM}

\item{languageHint}{Suggested backend language}

\item{streamingWorkerMainClass}{Streaming worker main class name}
}
\value{
TaskRunnerSettings object
}
\description{
TaskRunnerSettings Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Taskrunner configuration settings.
}

