% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{TopologyConfig.userStageToComputationNameMap}
\alias{TopologyConfig.userStageToComputationNameMap}
\title{TopologyConfig.userStageToComputationNameMap Object}
\usage{
TopologyConfig.userStageToComputationNameMap()
}
\value{
TopologyConfig.userStageToComputationNameMap object
}
\description{
TopologyConfig.userStageToComputationNameMap Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Maps user stage names to stable computation names.
}
\seealso{
Other TopologyConfig functions: \code{\link{TopologyConfig}}
}

