% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{TransformSummary}
\alias{TransformSummary}
\title{TransformSummary Object}
\usage{
TransformSummary(inputCollectionName = NULL, name = NULL, id = NULL,
  displayData = NULL, outputCollectionName = NULL)
}
\arguments{
\item{inputCollectionName}{User names for all collection inputs to this transform}

\item{name}{User provided name for this transform instance}

\item{id}{SDK generated id of this transform instance}

\item{displayData}{Transform-specific display data}

\item{outputCollectionName}{User  names for all collection outputs to this transform}
}
\value{
TransformSummary object
}
\description{
TransformSummary Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Description of the type, names/ids, and input/outputs for a transform.
}
