% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{WorkerHealthReport}
\alias{WorkerHealthReport}
\title{WorkerHealthReport Object}
\usage{
WorkerHealthReport(vmIsHealthy = NULL, vmStartupTime = NULL,
  reportInterval = NULL, pods = NULL)
}
\arguments{
\item{vmIsHealthy}{Whether the VM is healthy}

\item{vmStartupTime}{The time the VM was booted}

\item{reportInterval}{The interval at which the worker is sending health reports}

\item{pods}{The pods running on the worker}
}
\value{
WorkerHealthReport object
}
\description{
WorkerHealthReport Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
WorkerHealthReport contains information about the health of a worker. The VM should be identified by the labels attached to the WorkerMessage that this health ping belongs to.
}

