% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{WorkerHealthReportResponse}
\alias{WorkerHealthReportResponse}
\title{WorkerHealthReportResponse Object}
\usage{
WorkerHealthReportResponse(reportInterval = NULL)
}
\arguments{
\item{reportInterval}{A positive value indicates the worker should change its reporting interval to the specified value}
}
\value{
WorkerHealthReportResponse object
}
\description{
WorkerHealthReportResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
WorkerHealthReportResponse contains information returned to the worker in response to a health ping.
}

