% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{WorkerMessage.labels}
\alias{WorkerMessage.labels}
\title{WorkerMessage.labels Object}
\usage{
WorkerMessage.labels()
}
\value{
WorkerMessage.labels object
}
\description{
WorkerMessage.labels Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Labels are used to group WorkerMessages. For example, a worker_message about a particular container might have the labels: { 'JOB_ID': '2015-04-22', 'WORKER_ID': 'wordcount-vm-2015…' 'CONTAINER_TYPE': 'worker', 'CONTAINER_ID': 'ac1234def'} Label tags typically correspond to Label enum values. However, for ease of development other strings can be used as tags. LABEL_UNSPECIFIED should not be used here.
}
\seealso{
Other WorkerMessage functions: \code{\link{WorkerMessage}}
}

