% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{WorkerMessage}
\alias{WorkerMessage}
\title{WorkerMessage Object}
\usage{
WorkerMessage(WorkerMessage.labels = NULL, labels = NULL, time = NULL,
  workerHealthReport = NULL, workerMessageCode = NULL)
}
\arguments{
\item{WorkerMessage.labels}{The \link{WorkerMessage.labels} object or list of objects}

\item{labels}{Labels are used to group WorkerMessages}

\item{time}{The timestamp of the worker_message}

\item{workerHealthReport}{The health of a worker}

\item{workerMessageCode}{A worker message code}
}
\value{
WorkerMessage object
}
\description{
WorkerMessage Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
WorkerMessage provides information to the backend about a worker.
}
\seealso{
Other WorkerMessage functions: \code{\link{WorkerMessage.labels}}
}

