% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{WorkerMessageCode.parameters}
\alias{WorkerMessageCode.parameters}
\title{WorkerMessageCode.parameters Object}
\usage{
WorkerMessageCode.parameters()
}
\value{
WorkerMessageCode.parameters object
}
\description{
WorkerMessageCode.parameters Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Parameters contains specific information about the code. This is a struct to allow parameters of different types. Examples: 1. For a 'HARNESS_STARTED' message parameters might provide the name of the worker and additional data like timing information. 2. For a 'GCS_DOWNLOAD_ERROR' parameters might contain fields listing the GCS objects being downloaded and fields containing errors. In general complex data structures should be avoided. If a worker needs to send a specific and complicated data structure then please consider defining a new proto and adding it to the data oneof in WorkerMessageResponse. Conventions: Parameters should only be used for information that isn't typically passed as a label. hostname and other worker identifiers should almost always be passed as labels since they will be included on most messages.
}
\seealso{
Other WorkerMessageCode functions: \code{\link{WorkerMessageCode}}
}

