% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{WorkerMessageCode}
\alias{WorkerMessageCode}
\title{WorkerMessageCode Object}
\usage{
WorkerMessageCode(WorkerMessageCode.parameters = NULL, code = NULL,
  parameters = NULL)
}
\arguments{
\item{WorkerMessageCode.parameters}{The \link{WorkerMessageCode.parameters} object or list of objects}

\item{code}{The code is a string intended for consumption by a machine that identifies the type of message being sent}

\item{parameters}{Parameters contains specific information about the code}
}
\value{
WorkerMessageCode object
}
\description{
WorkerMessageCode Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A message code is used to report status and error messages to the service. The message codes are intended to be machine readable. The service will take care of translating these into user understandable messages if necessary. Example use cases: 1. Worker processes reporting successful startup. 2. Worker processes reporting specific errors (e.g. package staging failure).
}
\seealso{
Other WorkerMessageCode functions: \code{\link{WorkerMessageCode.parameters}}
}

