% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{WorkerMessageResponse}
\alias{WorkerMessageResponse}
\title{WorkerMessageResponse Object}
\usage{
WorkerMessageResponse(workerHealthReportResponse = NULL)
}
\arguments{
\item{workerHealthReportResponse}{The service's response to a worker's health report}
}
\value{
WorkerMessageResponse object
}
\description{
WorkerMessageResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A worker_message response allows the server to pass information to the sender.
}

