% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{WorkerPool.metadata}
\alias{WorkerPool.metadata}
\title{WorkerPool.metadata Object}
\usage{
WorkerPool.metadata()
}
\value{
WorkerPool.metadata object
}
\description{
WorkerPool.metadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Metadata to set on the Google Compute Engine VMs.
}
\seealso{
Other WorkerPool functions: \code{\link{WorkerPool.poolArgs}},
  \code{\link{WorkerPool}}
}
