% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{WorkerPool.poolArgs}
\alias{WorkerPool.poolArgs}
\title{WorkerPool.poolArgs Object}
\usage{
WorkerPool.poolArgs()
}
\value{
WorkerPool.poolArgs object
}
\description{
WorkerPool.poolArgs Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Extra arguments for this worker pool.
}
\seealso{
Other WorkerPool functions: \code{\link{WorkerPool.metadata}},
  \code{\link{WorkerPool}}
}

