% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{WorkItem}
\alias{WorkItem}
\title{WorkItem Object}
\usage{
WorkItem(packages = NULL, projectId = NULL, streamingSetupTask = NULL,
  sourceOperationTask = NULL, reportStatusInterval = NULL,
  leaseExpireTime = NULL, streamingConfigTask = NULL,
  initialReportIndex = NULL, streamingComputationTask = NULL,
  shellTask = NULL, jobId = NULL, id = NULL, configuration = NULL,
  mapTask = NULL, seqMapTask = NULL)
}
\arguments{
\item{packages}{Any required packages that need to be fetched in order to execute}

\item{projectId}{Identifies the cloud project this WorkItem belongs to}

\item{streamingSetupTask}{Additional information for StreamingSetupTask WorkItems}

\item{sourceOperationTask}{Additional information for source operation WorkItems}

\item{reportStatusInterval}{Recommended reporting interval}

\item{leaseExpireTime}{Time when the lease on this Work will expire}

\item{streamingConfigTask}{Additional information for StreamingConfigTask WorkItems}

\item{initialReportIndex}{The initial index to use when reporting the status of the WorkItem}

\item{streamingComputationTask}{Additional information for StreamingComputationTask WorkItems}

\item{shellTask}{Additional information for ShellTask WorkItems}

\item{jobId}{Identifies the workflow job this WorkItem belongs to}

\item{id}{Identifies this WorkItem}

\item{configuration}{Work item-specific configuration as an opaque blob}

\item{mapTask}{Additional information for MapTask WorkItems}

\item{seqMapTask}{Additional information for SeqMapTask WorkItems}
}
\value{
WorkItem object
}
\description{
WorkItem Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
WorkItem represents basic information about a WorkItem to be executedin the cloud.
}
