% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{WorkItemServiceState.harnessData}
\alias{WorkItemServiceState.harnessData}
\title{WorkItemServiceState.harnessData Object}
\usage{
WorkItemServiceState.harnessData()
}
\value{
WorkItemServiceState.harnessData object
}
\description{
WorkItemServiceState.harnessData Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Other data returned by the service, specific to the particularworker harness.
}
\seealso{
Other WorkItemServiceState functions: \code{\link{WorkItemServiceState}}
}
