% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{WorkItemServiceState}
\alias{WorkItemServiceState}
\title{WorkItemServiceState Object}
\usage{
WorkItemServiceState(WorkItemServiceState.harnessData = NULL,
  splitRequest = NULL, leaseExpireTime = NULL,
  reportStatusInterval = NULL, harnessData = NULL, nextReportIndex = NULL,
  metricShortId = NULL, suggestedStopPosition = NULL,
  suggestedStopPoint = NULL)
}
\arguments{
\item{WorkItemServiceState.harnessData}{The \link{WorkItemServiceState.harnessData} object or list of objects}

\item{splitRequest}{The progress point in the WorkItem where the Dataflow service suggests that the worker truncate the task}

\item{leaseExpireTime}{Time at which the current lease will expire}

\item{reportStatusInterval}{New recommended reporting interval}

\item{harnessData}{Other data returned by the service, specific to the particular worker harness}

\item{nextReportIndex}{The index value to use for the next report sent by the worker}

\item{metricShortId}{The short ids that workers should use in subsequent metric updates}

\item{suggestedStopPosition}{Obsolete, always empty}

\item{suggestedStopPoint}{DEPRECATED in favor of split_request}
}
\value{
WorkItemServiceState object
}
\description{
WorkItemServiceState Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The Dataflow service's idea of the current state of a WorkItem being processed by a worker.
}
\seealso{
Other WorkItemServiceState functions: \code{\link{WorkItemServiceState.harnessData}}
}

