% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{WorkItemStatus}
\alias{WorkItemStatus}
\title{WorkItemStatus Object}
\usage{
WorkItemStatus(counterUpdates = NULL, workItemId = NULL, errors = NULL,
  metricUpdates = NULL, dynamicSourceSplit = NULL,
  sourceOperationResponse = NULL, progress = NULL,
  requestedLeaseDuration = NULL, reportIndex = NULL, stopPosition = NULL,
  completed = NULL, reportedProgress = NULL, sourceFork = NULL)
}
\arguments{
\item{counterUpdates}{Worker output counters for this WorkItem}

\item{workItemId}{Identifies the WorkItem}

\item{errors}{Specifies errors which occurred during processing}

\item{metricUpdates}{DEPRECATED in favor of counter_updates}

\item{dynamicSourceSplit}{See documentation of stop_position}

\item{sourceOperationResponse}{If the work item represented a SourceOperationRequest, and the work}

\item{progress}{DEPRECATED in favor of reported_progress}

\item{requestedLeaseDuration}{Amount of time the worker requests for its lease}

\item{reportIndex}{The report index}

\item{stopPosition}{A worker may split an active map task in two parts, 'primary' and}

\item{completed}{True if the WorkItem was completed (successfully or unsuccessfully)}

\item{reportedProgress}{The worker's progress through this WorkItem}

\item{sourceFork}{DEPRECATED in favor of dynamic_source_split}
}
\value{
WorkItemStatus object
}
\description{
WorkItemStatus Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Conveys a worker's progress through the work described by a WorkItem.
}
