% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{WriteInstruction}
\alias{WriteInstruction}
\title{WriteInstruction Object}
\usage{
WriteInstruction(input = NULL, sink = NULL)
}
\arguments{
\item{input}{The input}

\item{sink}{The sink to write to}
}
\value{
WriteInstruction object
}
\description{
WriteInstruction Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An instruction that writes records.Takes one input, produces no outputs.
}
