% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{ComputationTopology}
\alias{ComputationTopology}
\title{ComputationTopology Object}
\usage{
ComputationTopology(systemStageName = NULL, computationId = NULL,
  userStageName = NULL, keyRanges = NULL, inputs = NULL, outputs = NULL,
  stateFamilies = NULL)
}
\arguments{
\item{systemStageName}{The system stage name}

\item{computationId}{The ID of the computation}

\item{userStageName}{The user stage name}

\item{keyRanges}{The key ranges processed by the computation}

\item{inputs}{The inputs to the computation}

\item{outputs}{The outputs from the computation}

\item{stateFamilies}{The state family values}
}
\value{
ComputationTopology object
}
\description{
ComputationTopology Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
All configuration data for a particular Computation.
}

