% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{ConcatPosition}
\alias{ConcatPosition}
\title{ConcatPosition Object}
\usage{
ConcatPosition(index = NULL, position = NULL)
}
\arguments{
\item{index}{Index of the inner source}

\item{position}{Position within the inner source}
}
\value{
ConcatPosition object
}
\description{
ConcatPosition Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A position that encapsulates an inner position and an index for the inner position. A ConcatPosition can be used by a reader of a source that encapsulates a set of other sources.
}

