% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{CounterStructuredName}
\alias{CounterStructuredName}
\title{CounterStructuredName Object}
\usage{
CounterStructuredName(originalStepName = NULL, workerId = NULL,
  originNamespace = NULL, origin = NULL, name = NULL,
  executionStepName = NULL, componentStepName = NULL, portion = NULL)
}
\arguments{
\item{originalStepName}{System generated name of the original step in the user's graph, before}

\item{workerId}{ID of a particular worker}

\item{originNamespace}{A string containing a more specific namespace of the counter's origin}

\item{origin}{One of the standard Origins defined above}

\item{name}{Counter name}

\item{executionStepName}{Name of the stage}

\item{componentStepName}{Name of the optimized step being executed by the workers}

\item{portion}{Portion of this counter, either key or value}
}
\value{
CounterStructuredName object
}
\description{
CounterStructuredName Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Identifies a counter within a per-job namespace. Counters whose structurednames are the same get merged into a single value for the job.
}
