% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{CreateJobFromTemplateRequest}
\alias{CreateJobFromTemplateRequest}
\title{CreateJobFromTemplateRequest Object}
\usage{
CreateJobFromTemplateRequest(CreateJobFromTemplateRequest.parameters = NULL,
  parameters = NULL, jobName = NULL, gcsPath = NULL, environment = NULL)
}
\arguments{
\item{CreateJobFromTemplateRequest.parameters}{The \link{CreateJobFromTemplateRequest.parameters} object or list of objects}

\item{parameters}{The runtime parameters to pass to the job}

\item{jobName}{Required}

\item{gcsPath}{Required}

\item{environment}{The runtime environment for the job}
}
\value{
CreateJobFromTemplateRequest object
}
\description{
CreateJobFromTemplateRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A request to create a Cloud Dataflow job from a template.
}
\seealso{
Other CreateJobFromTemplateRequest functions: \code{\link{CreateJobFromTemplateRequest.parameters}}
}
