% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{Environment.sdkPipelineOptions}
\alias{Environment.sdkPipelineOptions}
\title{Environment.sdkPipelineOptions Object}
\usage{
Environment.sdkPipelineOptions()
}
\value{
Environment.sdkPipelineOptions object
}
\description{
Environment.sdkPipelineOptions Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The Cloud Dataflow SDK pipeline options specified by the user. Theseoptions are passed through the service and are used to recreate theSDK pipeline options on the worker in a language agnostic and platformindependent way.
}
\seealso{
Other Environment functions: \code{\link{Environment.internalExperiments}},
  \code{\link{Environment.userAgent}},
  \code{\link{Environment.version}},
  \code{\link{Environment}}
}
