% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{GetDebugConfigRequest}
\alias{GetDebugConfigRequest}
\title{GetDebugConfigRequest Object}
\usage{
GetDebugConfigRequest(componentId = NULL, workerId = NULL)
}
\arguments{
\item{componentId}{The internal component id for which debug configuration is}

\item{workerId}{The worker id, i}
}
\value{
GetDebugConfigRequest object
}
\description{
GetDebugConfigRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request to get updated debug configuration for component.
}
