% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{KeyRangeDataDiskAssignment}
\alias{KeyRangeDataDiskAssignment}
\title{KeyRangeDataDiskAssignment Object}
\usage{
KeyRangeDataDiskAssignment(start = NULL, dataDisk = NULL, end = NULL)
}
\arguments{
\item{start}{The start (inclusive) of the key range}

\item{dataDisk}{The name of the data disk where data for this range is stored}

\item{end}{The end (exclusive) of the key range}
}
\value{
KeyRangeDataDiskAssignment object
}
\description{
KeyRangeDataDiskAssignment Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Data disk assignment information for a specific key-range of a shardedcomputation.Currently we only support UTF-8 character splits to simplify encoding intoJSON.
}
