% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{LeaseWorkItemRequest}
\alias{LeaseWorkItemRequest}
\title{LeaseWorkItemRequest Object}
\usage{
LeaseWorkItemRequest(currentWorkerTime = NULL, location = NULL,
  workItemTypes = NULL, workerCapabilities = NULL, workerId = NULL,
  requestedLeaseDuration = NULL)
}
\arguments{
\item{currentWorkerTime}{The current timestamp at the worker}

\item{location}{The location which contains the WorkItem's job}

\item{workItemTypes}{Filter for WorkItem type}

\item{workerCapabilities}{Worker capabilities}

\item{workerId}{Identifies the worker leasing work -- typically the ID of the}

\item{requestedLeaseDuration}{The initial lease period}
}
\value{
LeaseWorkItemRequest object
}
\description{
LeaseWorkItemRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request to lease WorkItems.
}
