% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{MapTask}
\alias{MapTask}
\title{MapTask Object}
\usage{
MapTask(systemName = NULL, stageName = NULL, instructions = NULL)
}
\arguments{
\item{systemName}{System-defined name of this MapTask}

\item{stageName}{System-defined name of the stage containing this MapTask}

\item{instructions}{The instructions in the MapTask}
}
\value{
MapTask object
}
\description{
MapTask Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
MapTask consists of an ordered set of instructions, each of whichdescribes one particular low-level operation for the worker toperform in order to accomplish the MapTask's WorkItem.Each instruction must appear in the list before any instructions whichdepends on its output.
}
