% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{Package}
\alias{Package}
\title{Package Object}
\usage{
Package(name = NULL, location = NULL)
}
\arguments{
\item{name}{The name of the package}

\item{location}{The resource to read the package from}
}
\value{
Package object
}
\description{
Package Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Packages that need to be installed in order for a worker to run the steps of the Dataflow job which will be assigned to its worker pool. This is the mechanism by which the SDK causes code to be loaded onto the workers. For example, the Dataflow Java SDK might use this to install jars containing the user's code and all of the various dependencies (libraries, data files, etc) required in order for that code to run.
}

