% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{Position}
\alias{Position}
\title{Position Object}
\usage{
Position(key = NULL, recordIndex = NULL, shufflePosition = NULL,
  byteOffset = NULL, concatPosition = NULL, end = NULL)
}
\arguments{
\item{key}{Position is a string key, ordered lexicographically}

\item{recordIndex}{Position is a record index}

\item{shufflePosition}{CloudPosition is a base64 encoded BatchShufflePosition (with FIXED}

\item{byteOffset}{Position is a byte offset}

\item{concatPosition}{CloudPosition is a concat position}

\item{end}{Position is past all other positions}
}
\value{
Position object
}
\description{
Position Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Position defines a position within a collection of data.  The valuecan be either the end position, a key (used with orderedcollections), a byte offset, or a record index.
}
