% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{PubsubLocation}
\alias{PubsubLocation}
\title{PubsubLocation Object}
\usage{
PubsubLocation(topic = NULL, subscription = NULL, timestampLabel = NULL,
  idLabel = NULL, dropLateData = NULL, trackingSubscription = NULL)
}
\arguments{
\item{topic}{A pubsub topic, in the form of 'pubsub}

\item{subscription}{A pubsub subscription, in the form of 'pubsub}

\item{timestampLabel}{If set, contains a pubsub label from which to extract record timestamps}

\item{idLabel}{If set, contains a pubsub label from which to extract record ids}

\item{dropLateData}{Indicates whether the pipeline allows late-arriving data}

\item{trackingSubscription}{If set, specifies the pubsub subscription that will be used for tracking custom time timestamps for watermark estimation}
}
\value{
PubsubLocation object
}
\description{
PubsubLocation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Identifies a pubsub location to use for transferring data into or out of a streaming Dataflow job.
}

