% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{ReportWorkItemStatusRequest}
\alias{ReportWorkItemStatusRequest}
\title{ReportWorkItemStatusRequest Object}
\usage{
ReportWorkItemStatusRequest(workerId = NULL, workItemStatuses = NULL,
  currentWorkerTime = NULL)
}
\arguments{
\item{workerId}{The ID of the worker reporting the WorkItem status}

\item{workItemStatuses}{The order is unimportant, except that the order of the WorkItemServiceState messages in the ReportWorkItemStatusResponse corresponds to the order of WorkItemStatus messages here}

\item{currentWorkerTime}{The current timestamp at the worker}
}
\value{
ReportWorkItemStatusRequest object
}
\description{
ReportWorkItemStatusRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request to report the status of WorkItems.
}

