% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{SeqMapTaskOutputInfo}
\alias{SeqMapTaskOutputInfo}
\title{SeqMapTaskOutputInfo Object}
\usage{
SeqMapTaskOutputInfo(sink = NULL, tag = NULL)
}
\arguments{
\item{sink}{The sink to write the output value to}

\item{tag}{The id of the TupleTag the user code will tag the output value by}
}
\value{
SeqMapTaskOutputInfo object
}
\description{
SeqMapTaskOutputInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Information about an output of a SeqMapTask.
}
