% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{Source}
\alias{Source}
\title{Source Object}
\usage{
Source(Source.spec = NULL, Source.codec = NULL, spec = NULL,
  codec = NULL, baseSpecs = NULL, metadata = NULL,
  doesNotNeedSplitting = NULL)
}
\arguments{
\item{Source.spec}{The \link{Source.spec} object or list of objects}

\item{Source.codec}{The \link{Source.codec} object or list of objects}

\item{spec}{The source to read from, plus its parameters}

\item{codec}{The codec to use to decode data read from the source}

\item{baseSpecs}{While splitting, sources may specify the produced bundles as differences against another source, in order to save backend-side memory and allow bigger jobs}

\item{metadata}{Optionally, metadata for this source can be supplied right away, avoiding a SourceGetMetadataOperation roundtrip (see SourceOperationRequest)}

\item{doesNotNeedSplitting}{Setting this value to true hints to the framework that the source doesn't need splitting, and using SourceSplitRequest on it would yield SOURCE_SPLIT_OUTCOME_USE_CURRENT}
}
\value{
Source object
}
\description{
Source Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A source that records can be read and decoded from.
}
\seealso{
Other Source functions: \code{\link{Source.codec}},
  \code{\link{Source.spec}}
}

