% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{SourceFork}
\alias{SourceFork}
\title{SourceFork Object}
\usage{
SourceFork(residual = NULL, residualSource = NULL, primary = NULL,
  primarySource = NULL)
}
\arguments{
\item{residual}{DEPRECATED}

\item{residualSource}{DEPRECATED}

\item{primary}{DEPRECATED}

\item{primarySource}{DEPRECATED}
}
\value{
SourceFork object
}
\description{
SourceFork Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
DEPRECATED in favor of DynamicSourceSplit.
}
