% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{SourceOperationRequest}
\alias{SourceOperationRequest}
\title{SourceOperationRequest Object}
\usage{
SourceOperationRequest(split = NULL, getMetadata = NULL)
}
\arguments{
\item{split}{Information about a request to split a source}

\item{getMetadata}{Information about a request to get metadata about a source}
}
\value{
SourceOperationRequest object
}
\description{
SourceOperationRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A work item that represents the different operations that can be performed on a user-defined Source specification.
}

