% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{SourceSplitShard}
\alias{SourceSplitShard}
\title{SourceSplitShard Object}
\usage{
SourceSplitShard(source = NULL, derivationMode = NULL)
}
\arguments{
\item{source}{DEPRECATED}

\item{derivationMode}{DEPRECATED}
}
\value{
SourceSplitShard object
}
\description{
SourceSplitShard Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
DEPRECATED in favor of DerivedSource.
}

