% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{StreamingSideInputLocation}
\alias{StreamingSideInputLocation}
\title{StreamingSideInputLocation Object}
\usage{
StreamingSideInputLocation(stateFamily = NULL, tag = NULL)
}
\arguments{
\item{stateFamily}{Identifies the state family where this side input is stored}

\item{tag}{Identifies the particular side input within the streaming Dataflow job}
}
\value{
StreamingSideInputLocation object
}
\description{
StreamingSideInputLocation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Identifies the location of a streaming side input.
}
