% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{StreamLocation}
\alias{StreamLocation}
\title{StreamLocation Object}
\usage{
StreamLocation(streamingStageLocation = NULL, pubsubLocation = NULL,
  sideInputLocation = NULL, customSourceLocation = NULL)
}
\arguments{
\item{streamingStageLocation}{The stream is part of another computation within the current streaming Dataflow job}

\item{pubsubLocation}{The stream is a pubsub stream}

\item{sideInputLocation}{The stream is a streaming side input}

\item{customSourceLocation}{The stream is a custom source}
}
\value{
StreamLocation object
}
\description{
StreamLocation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Describes a stream of data, either as input to be processed or as output of a streaming Dataflow job.
}

