% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{TaskRunnerSettings}
\alias{TaskRunnerSettings}
\title{TaskRunnerSettings Object}
\usage{
TaskRunnerSettings(workflowFileName = NULL, tempStoragePrefix = NULL,
  baseTaskDir = NULL, commandlinesFileName = NULL, languageHint = NULL,
  baseUrl = NULL, logToSerialconsole = NULL, continueOnException = NULL,
  parallelWorkerSettings = NULL, vmId = NULL, taskUser = NULL,
  alsologtostderr = NULL, taskGroup = NULL, harnessCommand = NULL,
  logDir = NULL, oauthScopes = NULL, dataflowApiVersion = NULL,
  logUploadLocation = NULL, streamingWorkerMainClass = NULL)
}
\arguments{
\item{workflowFileName}{The file to store the workflow in}

\item{tempStoragePrefix}{The prefix of the resources the taskrunner should use for}

\item{baseTaskDir}{The location on the worker for task-specific subdirectories}

\item{commandlinesFileName}{The file to store preprocessing commands in}

\item{languageHint}{The suggested backend language}

\item{baseUrl}{The base URL for the taskrunner to use when accessing Google Cloud APIs}

\item{logToSerialconsole}{Whether to send taskrunner log info to Google Compute Engine VM serial}

\item{continueOnException}{Whether to continue taskrunner if an exception is hit}

\item{parallelWorkerSettings}{The settings to pass to the parallel worker harness}

\item{vmId}{The ID string of the VM}

\item{taskUser}{The UNIX user ID on the worker VM to use for tasks launched by}

\item{alsologtostderr}{Whether to also send taskrunner log info to stderr}

\item{taskGroup}{The UNIX group ID on the worker VM to use for tasks launched by}

\item{harnessCommand}{The command to launch the worker harness}

\item{logDir}{The directory on the VM to store logs}

\item{oauthScopes}{The OAuth2 scopes to be requested by the taskrunner in order to}

\item{dataflowApiVersion}{The API version of endpoint, e}

\item{logUploadLocation}{Indicates where to put logs}

\item{streamingWorkerMainClass}{The streaming worker main class name}
}
\value{
TaskRunnerSettings object
}
\description{
TaskRunnerSettings Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Taskrunner configuration settings.
}
