% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{WorkerHealthReport}
\alias{WorkerHealthReport}
\title{WorkerHealthReport Object}
\usage{
WorkerHealthReport(pods = NULL, vmStartupTime = NULL, vmIsHealthy = NULL,
  reportInterval = NULL)
}
\arguments{
\item{pods}{The pods running on the worker}

\item{vmStartupTime}{The time the VM was booted}

\item{vmIsHealthy}{Whether the VM is healthy}

\item{reportInterval}{The interval at which the worker is sending health reports}
}
\value{
WorkerHealthReport object
}
\description{
WorkerHealthReport Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
WorkerHealthReport contains information about the health of a worker.The VM should be identified by the labels attached to the WorkerMessage thatthis health ping belongs to.
}
