% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{WorkerSettings}
\alias{WorkerSettings}
\title{WorkerSettings Object}
\usage{
WorkerSettings(servicePath = NULL, shuffleServicePath = NULL,
  workerId = NULL, tempStoragePrefix = NULL, reportingEnabled = NULL,
  baseUrl = NULL)
}
\arguments{
\item{servicePath}{The Cloud Dataflow service path relative to the root URL, for example,}

\item{shuffleServicePath}{The Shuffle service path relative to the root URL, for example,}

\item{workerId}{The ID of the worker running this pipeline}

\item{tempStoragePrefix}{The prefix of the resources the system should use for temporary}

\item{reportingEnabled}{Whether to send work progress updates to the service}

\item{baseUrl}{The base URL for accessing Google Cloud APIs}
}
\value{
WorkerSettings object
}
\description{
WorkerSettings Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Provides data to pass through to the worker harness.
}
