% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{WorkItem}
\alias{WorkItem}
\title{WorkItem Object}
\usage{
WorkItem(id = NULL, projectId = NULL, jobId = NULL, packages = NULL,
  mapTask = NULL, seqMapTask = NULL, shellTask = NULL,
  streamingSetupTask = NULL, sourceOperationTask = NULL,
  streamingComputationTask = NULL, streamingConfigTask = NULL,
  reportStatusInterval = NULL, leaseExpireTime = NULL,
  configuration = NULL, initialReportIndex = NULL)
}
\arguments{
\item{id}{Identifies this WorkItem}

\item{projectId}{Identifies the cloud project this WorkItem belongs to}

\item{jobId}{Identifies the workflow job this WorkItem belongs to}

\item{packages}{Any required packages that need to be fetched in order to execute this WorkItem}

\item{mapTask}{Additional information for MapTask WorkItems}

\item{seqMapTask}{Additional information for SeqMapTask WorkItems}

\item{shellTask}{Additional information for ShellTask WorkItems}

\item{streamingSetupTask}{Additional information for StreamingSetupTask WorkItems}

\item{sourceOperationTask}{Additional information for source operation WorkItems}

\item{streamingComputationTask}{Additional information for StreamingComputationTask WorkItems}

\item{streamingConfigTask}{Additional information for StreamingConfigTask WorkItems}

\item{reportStatusInterval}{Recommended reporting interval}

\item{leaseExpireTime}{Time when the lease on this Work will expire}

\item{configuration}{Work item-specific configuration as an opaque blob}

\item{initialReportIndex}{The initial index to use when reporting the status of the WorkItem}
}
\value{
WorkItem object
}
\description{
WorkItem Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
WorkItem represents basic information about a WorkItem to be executed in the cloud.
}

