% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_objects.R
\name{WorkItemStatus}
\alias{WorkItemStatus}
\title{WorkItemStatus Object}
\usage{
WorkItemStatus(workItemId = NULL, reportIndex = NULL,
  requestedLeaseDuration = NULL, completed = NULL, errors = NULL,
  counterUpdates = NULL, metricUpdates = NULL, reportedProgress = NULL,
  stopPosition = NULL, dynamicSourceSplit = NULL,
  sourceOperationResponse = NULL, sourceFork = NULL, progress = NULL)
}
\arguments{
\item{workItemId}{Identifies the WorkItem}

\item{reportIndex}{The report index}

\item{requestedLeaseDuration}{Amount of time the worker requests for its lease}

\item{completed}{True if the WorkItem was completed (successfully or unsuccessfully)}

\item{errors}{Specifies errors which occurred during processing}

\item{counterUpdates}{Worker output counters for this WorkItem}

\item{metricUpdates}{DEPRECATED in favor of counter_updates}

\item{reportedProgress}{The worker's progress through this WorkItem}

\item{stopPosition}{A worker may split an active map task in two parts, 'primary' and 'residual', continuing to process the primary part and returning the residual part into the pool of available work}

\item{dynamicSourceSplit}{See documentation of stop_position}

\item{sourceOperationResponse}{If the work item represented a SourceOperationRequest, and the work is completed, contains the result of the operation}

\item{sourceFork}{DEPRECATED in favor of dynamic_source_split}

\item{progress}{DEPRECATED in favor of reported_progress}
}
\value{
WorkItemStatus object
}
\description{
WorkItemStatus Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Conveys a worker's progress through the work described by a WorkItem.
}

